# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/file/upload": {
      "post": {
        "operationId": "FileController_uploadFile_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "File upload",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload file successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "asset_id": "some-asset-id",
                      "public_id": "some-public-id",
                      "url": "https://res.cloudinary.com/..."
                    },
                    "message": "File uploaded successfully"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload file",
        "tags": [
          "File"
        ]
      }
    },
    "/api/v1/file/upload-multiple": {
      "post": {
        "operationId": "FileController_uploadMultipleFiles_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Multiple file upload",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Upload multiple files successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": [
                      {
                        "asset_id": "some-asset-id",
                        "public_id": "some-public-id",
                        "url": "https://res.cloudinary.com/..."
                      }
                    ],
                    "message": "Files uploaded successfully"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload multiple files",
        "tags": [
          "File"
        ]
      }
    },
    "/api/v1/file/delete-multiple": {
      "delete": {
        "operationId": "FileController_deleteMultipleFiles_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Delete multiple files by URLs",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteFilesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Delete multiple files successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": [
                      {
                        "result": "ok",
                        "deleted": {
                          "folder/filename": "deleted"
                        }
                      }
                    ],
                    "message": "Files deleted successfully"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete multiple files by file URLs",
        "tags": [
          "File"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "Swagger docs for the project",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DeleteFilesDto": {
        "type": "object",
        "properties": {
          "urls": {
            "description": "Array of Cloudinary URLs to delete",
            "example": [
              "https://res.cloudinary.com/cloud_name/image/upload/v1234567890/folder/filename.jpg",
              "https://res.cloudinary.com/cloud_name/video/upload/v1234567890/folder/video.mp4"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "urls"
        ]
      },
...
    }
  }
}